agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent"
version = "0.1.0"
description = "A science chat agent for distilling academic papers."
dependencies = [
    "models",
    "aiohttp",
    "aiofiles",
    "jinja2",
    "python-dotenv",
    "fastapi",
    "uvicorn[standard]",
    "prompt_toolkit"
]

[project.scripts]
agent-template = "agent.main:main"

[tool.uv.sources]
scope_loader = { path = "../scope-loader" }
taskman = { path = "../taskman" }
models = { path = "../models" }

[tool.hatch.build.targets.wheel]
packages = ["agent"]
